Create TestEPS_Scenario.java
TestEPS_Scenario is mean to have a schedule, and inplement a way to talk to TENMA72-2540.

Currently it can olny change the voltage on TENMA72-2540.

Uncommenting the try, will obtain housekeeping data from the like TestEPS.java

Also removed the BooleanExpression from the EPS.xml. The reason for that is that it failed to build. However the program works without the BooleanExpression.